requirements.txt
django-cors-headers==3.2.1